+++ NEWS.md
@@ -1,6 +1,7 @@
 # brandseyer2 0.0.1.9006
 
 * bug fix: `mentions()` no longer raises a condition if `id` isn't selected.
+* bug fix: `root_brands()`, when operating on a list of accounts, now includes the account code.
 
 # brandseyer2 0.0.1.9005